Improve Content Browser, Blink, Base, UI Events
We now build the "child" part of Content Browser. It
implements most of the Blink logic.

Thus, we also build the required "webcrypto" component
in Content, which we patch back to NSS, and thus depend
on "nss" as a result.

We now generate Grit resources in Blink, which are
also required.

We now build some Blink-specific Gesture source in
UI Events Blink, which are also required.

We add required debug profiler support in Base (which means
building "debug/profiler.cc", which requires file generation,
itself requiring Python.

Signed-off-by: Manuel Bachmann <mbc@iot.bzh>