Add optional Fontconfig, rename FreeType shared library
If we want to use the optional FreeType to really NOT depend
on system FreeType, we also want optional Fontconfig because
system Fontconfig depends on system Freetype and pulls it
indirectly.

Rename the FreeType shared object from "libfreetype.so" to
"libcrfreetype.so", so that system "libfreetype.so" does not
mistakenly get picked if development packages are present.

Signed-off-by: Manuel Bachmann <mbc@iot.bzh>